Add files via upload
Developed a machine learning-based web application using Streamlit that predicts the job category of uploaded resumes. It uses natural language processing to clean and vectorize resume text and a trained classifier to match resumes to one of 25 career categories. Integrated a sleek UI with a custom background for better UX.